refactor: 프로젝트 라우팅 경로 정리
